Totally rework for Contao 4
